requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.23.2

# LLM and RAG
langchain==0.0.335
langchain-openai==0.0.2
openai==1.2.3
faiss-cpu==1.7.4
scikit-learn==1.3.0

# Data Processing
pandas==2.1.1
numpy==1.26.0
pydantic==2.4.2
python-dateutil==2.8.2

# API Clients
requests==2.31.0
python-dotenv==1.0.0
alpaca-trade-api==3.0.2
newsapi-python==0.2.7

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Portfolio Optimization
pyportfolioopt==1.5.5

# Visualization
matplotlib==3.8.0
seaborn==0.13.0

# Testing
pytest==7.4.3

# Utilities
tqdm==4.66.1
click==8.1.7
